find quotes from checkio
Your task at hand is to find all quotes in the text. And, as is already usual, do everything as quickly as possible :)
You are given a string that consists of characters and a paired number of quotation marks. You need to return an Iterable consisting of the texts inside the quotation marks. But choose only quotes with double quotation marks ("). Single quotation marks (') aren’t appropriate.

Input: A string.

Output: An iterable.

Example:

find_quotes('"Greetings"') == ['Greetings']
find_quotes('Hi') == []